requirements.txt
flet>=0.82.0
google-genai>=1.0.0
groq>=0.4.0
fpdf2>=2.8.0
requests>=2.31.0
